Merge 0.1 head

See merge request olena/pylene!97

+++ tests/accu/max_elements.cpp
@@ -36,23 +36,23 @@ TEST(Accu, Max_elements_vec)
 
   accumulators::maximal_elements<rgb8> acc;
 
-  acc.take(rgb8{(uint8)5, (uint8)5, (uint8)255});
-  acc.take(rgb8{(uint8)255, (uint8)5, (uint8)255});
-  acc.take(rgb8{(uint8)2, (uint8)2, (uint8)25});
+  acc.take(rgb8{5, 5, 255});
+  acc.take(rgb8{255, 5, 255});
+  acc.take(rgb8{2, 2, 25});
 
   {
     auto res = extractor::maximal_elements(acc);
     ASSERT_EQ(res.size(), 1u);
-    ASSERT_EQ(res[0], (rgb8{(uint8)255, (uint8)5, (uint8)255}));
+    ASSERT_EQ(res[0], (rgb8{255, 5, 255}));
   }
 
-  acc.take(rgb8{(uint8)5, (uint8)8, (uint8)50});
-  acc.take(rgb8{(uint8)4, (uint8)1, (uint8)25});
+  acc.take(rgb8{5, 8, 50});
+  acc.take(rgb8{4, 1, 25});
 
   {
     auto res = extractor::maximal_elements(acc);
     ASSERT_EQ(res.size(), 2u);
-    ASSERT_EQ(res[0], (rgb8{(uint8)255, (uint8)5, (uint8)255}));
-    ASSERT_EQ(res[1], (rgb8{(uint8)5, (uint8)8, (uint8)50}));
+    ASSERT_EQ(res[0], (rgb8{255, 5, 255}));
+    ASSERT_EQ(res[1], (rgb8{5, 8, 50}));
   }
 }
